@@ -12,31 +12,26 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "7.0.2",
-    "@angular/animations": "^7.1.4",
-    "@angular/cdk": "^7.1.4",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
+    "@angular/animations": "^8.0.0",
+    "@angular/cdk": "^8.0.0",
+    "@angular/common": "^8.0.0",
+    "@angular/compiler": "^8.0.0",
+    "@angular/core": "^8.0.0",
     "@angular/flex-layout": "7.0.0-beta.19",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
-    "@angular/material": "^7.1.4",
-    "@angular/material-moment-adapter": "^7.1.4",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/router": "^7.1.4",
-    "@swimlane/ngx-charts": "^7.0.1",
-    "@swimlane/ngx-datatable": "^11.1.5",
-    "angular-calendar": "0.25.2",
+    "@angular/forms": "^8.0.0",
+    "@angular/material": "^8.0.0",
+    "@angular/material-moment-adapter": "^8.0.0",
+    "@angular/platform-browser": "^8.0.0",
+    "@angular/platform-browser-dynamic": "^8.0.0",
+    "@angular/platform-server": "^8.0.0",
+    "@angular/router": "^8.0.0",
     "apollo-angular": "^1.4.1",
     "apollo-angular-link-http": "^1.3.1",
     "apollo-cache-inmemory": "^1.3.5",
     "apollo-client": "^2.4.2",
     "auth0-js": "^9.8.0",
     "bootstrap": "^4.1.3",
-    "chartist": "^0.11.0",
-    "core-js": "2.4.1",
+    "core-js": "^3.1.3",
     "d3": "^4.13.0",
     "d3-scale-chromatic": "^1.4.0",
     "express": "^4.16.4",
@@ -60,35 +55,36 @@
     "ngx-perfect-scrollbar": "6.1.0",
     "ngx-quill": "^3.1.0",
     "path": "^0.12.7",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.0",
     "rxjs-compat": "^6.2.0",
-    "rxjs-tslint": "^0.1.4",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.4",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
+    "@angular-devkit/build-angular": "^0.800.0",
+    "@angular/cli": "^8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/language-service": "^8.0.0",
     "@types/chartist": "0.9.37",
     "@types/d3": "^5.7.2",
-    "@types/jasmine": "~2.8.6",
+    "@types/jasmine": "^3.5.9",
     "@types/jasminewd2": "~2.0.3",
     "@types/jquery": "^3.3.22",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
+    "@types/node": "^13.9.1",
+    "codelyzer": "^5.2.1",
     "enhanced-resolve": "^3.3.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
+    "jasmine-core": "^3.5.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "^5.4.1",
+    "karma-jasmine": "^3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.2",
+    "protractor": "^5.4.3",
+    "rxjs-tslint": "^0.1.7",
     "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~3.1.6"
+    "tslint": "^6.1.0",
+    "typescript": "3.4.5",
+    "webpack": "^4.42.0"
   }
 }
