@@ -24,7 +24,7 @@ export class CollectionDialogComponent implements OnInit, AfterViewInit, OnDestr
     public filteredUsers: ReplaySubject<User[]> = new ReplaySubject<User[]>(1);
     protected _onDestroy = new Subject<void>();
 
-    @ViewChild('userSelect') userSelect: MatSelect;
+    @ViewChild('userSelect', {static: true}) userSelect: MatSelect;
 
     constructor(
         public dialogRef: MatDialogRef<CollectionDialogComponent>,
