@@ -22,8 +22,8 @@ export class DataSourceComponent implements OnInit, AfterViewInit  {
         }
     ];
 
-    @ViewChild(MatSort) sort: MatSort;
-    @ViewChild(MatPaginator) paginator: MatPaginator;
+    @ViewChild(MatSort, {static: true}) sort: MatSort;
+    @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
 
     constructor(private route: ActivatedRoute,
                 private dataSourceService: DataSourceService,
